requirements_dev.txt
sanic-testing
pytest-asyncio